move reaction generation functions to rmgpy.rmg.react module
